requirements.txt
autobahn==17.6.2
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.6.0
bitarray==0.8.1
Django==1.10.5
docker-py==1.10.3
docker-pycreds==0.2.1
fastzbarlight==0.0.14
funcsigs==1.0.2
gitdb2==2.0.2
GitPython==2.1.5
html5lib==0.999999999
incremental==17.5.0
ipaddress==1.0.18
mock==2.0.0
numpy==1.13.0
olefile==0.44
pbr==3.0.1
Pillow==4.2.0
protobuf==3.3.0
psycopg2==2.7.1
pygame==1.9.3
pyglet==1.2.4
pytz==2017.2
PyYAML==3.12
requests==2.10.0
scipy==0.19.0
simplejson==3.10.0
six==1.10.0
smmap2==2.0.3
txaio==2.8.0
virtualenv==15.1.0
webencodings==0.5.1
websocket-client==0.44.0
